Avoids XmpUtils.generate_xmp to overwrite files.

+++ SpiMediaGallery/main/xmp_utils.py
@@ -1,3 +1,4 @@
+import os
 import textwrap
 from typing import Set
 
@@ -7,6 +8,11 @@ from libxmp.utils import file_to_dict
 class XmpUtils:
     @staticmethod
     def read_tags(file_path: str) -> Set[str]:
+        """Reads tags from an XMP file (only Digikam tags)
+
+        :param file_path: File where to read the tags from
+        :return: Set of tags
+        """
         tags = set()
 
         xmp = file_to_dict(file_path)
@@ -24,10 +30,19 @@ class XmpUtils:
 
     @staticmethod
     def generate_xmp(output_file: str, tags: Set[str]) -> None:
+        """Generates a new XMP file with the set of tags. Only for Digikam tags
+
+        :param output_file: output file
+        :param tags: tags to write there
+        """
         # libxmp Python library seems to not be able to create a *new* XMP file, only
         # to open, read and modify files
         # xml Python library could do this but with some problems for the root node;
         # to have the finest control over the format the XML is hardcoded here
+
+        if os.path.exists(output_file):
+            raise FileExistsError("{} exists".format(output_file))
+
         with open(output_file, "w") as xmp_file:
             formatted_tags = ""
             for tag in tags:
